Convert from Map -> NodeIndex
Map's are fairly limited in browsers: they can only handle 4,000,000 entries. We will need to be able handle OSM data with at least 40,000,000 nodes and possibly more. Knowing this limitation is here and knowing that we will want to able to handle processing in a `WebWorker` has led me to create the `NodeIndex` class to handle parsing and retrieving `OsmNode`s where all values are stored in TypedArrays and a string lookup table (eventually, that will be shared externally).