Add entity tag key -> value search
Search for entities by tag keys and values. Stores a tag key -> entity look up table in the tag index. When searching for tag key=value pairs it finds all of the entities that have that tag key and then checks to see if it contains the corresponding value as well. If use case requires it, we can add more efficient key=value lookup tables at a later date.