Router (#28)
Build a routable graph using the Osmix core indexes that can be used to
generate routes across OSM networks.

Aims to be a simple implementation that can be used as a reference or
built on top of.

Currently does not take turn restrictions into account.

Example implementation is in the merge app on the inspect page. Click
two points to see a route.